server/requirements-dev.txt
# Include base requirements
-r requirements.txt

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0

# Code Quality
flake8==6.1.0
black==23.12.0
isort==5.13.2
mypy==1.7.1

# Development
python-dotenv==1.0.0
flask-shell-ipython==0.4.1
ipython==8.17.2
